refactor: create /Draft for all Draft.js, include DraftforNailCosmic.
