+++ app/Component/Draft/DraftDisplay.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import {Editor, EditorState,convertToRaw, convertFromRaw} from 'draft-js';
+import styles from "./styleDisplay.module.css";
 
 export default class DraftDisplay extends React.Component {
   constructor(props){
@@ -8,6 +9,7 @@ export default class DraftDisplay extends React.Component {
       editorState: this.props.editorState?EditorState.createWithContent(convertFromRaw(this.props.editorState)):EditorState.createEmpty(),
     };
     this.changeEditorState = (editorState) => {this.setState({editorState: editorState})};
+    this._draft_blockClass = this._draft_blockClass.bind(this);
     this.style={
       Com_DraftDisplay_: {
         width: '100%',
@@ -19,6 +21,10 @@ export default class DraftDisplay extends React.Component {
     }
   }
 
+  _draft_blockClass(contentBlock){
+    return styles.markDisplayBlock;
+  }
+
   render(){
     //let cx = cxBind.bind(styles);
     return(
@@ -28,6 +34,7 @@ export default class DraftDisplay extends React.Component {
           ref={(element)=>{this.contentEditor = element;}}
           editorState={this.state.editorState}
           onChange={this.changeEditorState}
+          blockStyleFn={this._draft_blockClass}
           readOnly/>
       </div>
     )
