+++ app/Component/Unit/Mark/MarkDialogue.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Editor, EditorState,convertToRaw, convertFromRaw} from 'draft-js';
-import DraftEditor from '../../DraftEditor.jsx';
-import DraftDisplay from '../../DraftDisplay.jsx';
+import DraftEditor from '../../Draft/DraftEditor.jsx';
+import DraftDisplay from '../../Draft/DraftDisplay.jsx';
 
 export default class MarkDialogue extends React.Component {
   constructor(props){
