+++ app/Within/component/RelationRes.jsx
@@ -8,7 +8,7 @@ import {connect} from "react-redux";
 import cxBind from 'classnames/bind';
 import Unit from '../../Component/Unit.jsx';
 import SvgPropic from '../../Component/Svg/SvgPropic.jsx';
-import DraftDisplay from '../../Component/DraftDisplay.jsx';
+import DraftDisplay from '../../Component/Draft/DraftDisplay.jsx';
 
 class RelationRes extends React.Component {
   constructor(props){
