+++ browserify/bunDevelope.js
@@ -32,9 +32,9 @@ let rootSelfFront = browserify({debug: true}).transform(babelify.configure({
   plugins: [
   	"transform-object-rest-spread"
   ]})
-).require("./app/Self/root_Front.js", {entry: true})/*.plugin(require('css-modulesify'), {
+).require("./app/Self/root_Front.js", {entry: true}).plugin(require('css-modulesify'), {
     rootDir: __dirname
-  });*/
+  });
 let rootTerrace = browserify({debug: true}).transform(babelify.configure({
   presets: [
       "react",
@@ -53,6 +53,9 @@ let appSelfFront = rootSelfFront.bundle().on("error", function (err) { console.l
 let appTerrace = rootTerrace.bundle().on("error", function (err) { console.log("Error: " + err.message); });
 rootWithin.on('css stream', function (css) {
     css.pipe(fs.createWriteStream('./public/css/stylesWithin.css')); //rewrite the file with the new "abstract name"
+});
+rootSelfFront.on('css stream', function (css) {
+    css.pipe(fs.createWriteStream('./public/css/stylesSelfFront.css')); //rewrite the file with the new "abstract name"
 });
 
 exports.bundler = ()=>{
