+++ public/css/stylesSelfFront.css
@@ -12,11 +12,3 @@
 ._C_C_app_Component_Draft_styleDisplay_module__markDisplayBlock div, ._C_C_app_Component_Draft_styleDisplay_module__markDisplayBlockNowrap div {
   display: inline-block;
 }
-._C_C_app_Component_Nails_NailCosmic_styleNailCosmic_module__Nails_Cosmic {
-  display: inline-block;
-  width: 100%;
-  position: relative;
-  box-sizing: border-box;
-  border-radius: 0.08vh;
-  cursor: pointer;
-}
