+++ public/html/html_SelfFront.html
@@ -14,6 +14,7 @@
     <link type='text/css' rel="stylesheet" href="https://fonts.googleapis.com/earlyaccess/cwtexming.css"/>
     <link href="https://fonts.googleapis.com/css?family=Lato:100,300,400&amp;subset=latin-ext" rel="stylesheet">
     <link type='text/css' rel="stylesheet" href="/css/styleSelfCommon.css"/>
+    <link type='text/css' rel="stylesheet" href="/css/stylesSelfFront.css"/>
     <style>
       html {font-size: 10px;}
       body {font-family: 'Lato', 'Noto Sans TC', '微軟正黑體', 'Helvetica Neue', Helvetica, Futura, sans-serif, Arial;}
